--- a/benchmarks/src/test/scala/org/scalafmt/BenchmarkOK.scala
+++ b/benchmarks/src/test/scala/org/scalafmt/BenchmarkOK.scala
@@ -10,10 +10,12 @@ class BenchmarkOK extends FunSuite with FormatAssertions {
 
   // TODO(olafur) DRY.
   val formatBenchmarks = Seq(
-    new OptimizerCore,
-    new SourceMapWriter,
-    new Division,
-    new BaseLinker
+    new GenJsCode,
+    new ScalaJSClassEmitter,
+    new OptimizerCore
+//    new SourceMapWriter,
+//    new Division,
+//    new BaseLinker
   )
   formatBenchmarks.foreach { formatBenchmark =>
     val name = formatBenchmark.getClass.getSimpleName