--- a/core/src/main/scala/org/scalafmt/internal/BestFirstSearch.scala
+++ b/core/src/main/scala/org/scalafmt/internal/BestFirstSearch.scala
@@ -26,39 +26,86 @@ class BestFirstSearch(style: ScalaStyle, tree: Tree, range: Set[Range])
 
   import BestFirstSearch._
 
-  val maxVisitStates = 10000000 // For debugging purposes only.
+  val maxVisitStates =  // For debugging purposes only.
+    if (style.debug) 100000 // Unit tests must be < 100k states
+    else 10000000
+
   val doOptimizations = true // For debugging purposes only.
-  val pruneNonOptimal = doOptimizations
-  val dequeOnNewStatements = doOptimizations
+
+  /**
+    * Eliminate solutions that are inferior to already explored solutions.
+    *
+    * If a solution is reaches a point at cost X then any other solution that
+    * reaches the same token with a cost > X will be eliminated.
+    */
+  val pruneNonOptimal = true && doOptimizations
+
+  /**
+    * When entering a new statement, clear out search queue.
+    */
+  val dequeOnNewStatements = true && doOptimizations
+
+  /**
+    * Recursively format {} wrapped blocks inside no optimization zones.
+    *
+    * By starting a new search queue, we can perform aggressive optimizations
+    * inside optimizations zones.
+    */
+  val recurseOnBlocks = true && doOptimizations
+
   val toks: Array[FormatToken] = FormatToken.formatTokens(tree.tokens)
   val owners = getOwners(tree)
   val statementStarts = getStatementStarts(tree)
+  val matchingParentheses = getMatchingParentheses(tree.tokens)
   val noOptimizations = noOptimizationZones(tree)
   val best = mutable.Map.empty[Token, State]
-  val matchingParentheses = getMatchingParentheses(tree.tokens)
   val router = new Router(style, tree, toks,
     matchingParentheses, statementStarts, owners)
   var explored = 0
   var deepestYet = State.start
   var statementCount = 0
 
+  type StateHash = Long
+
 
   def isInsideNoOptZone(token: FormatToken): Boolean = {
-    // TODO(olafur) use more efficient range map, for example Guava.
     noOptimizations.contains(token.left)
   }
+
   /**
     * Returns true if it's OK to skip over state.
     */
-  def pruneOK(curr: State): Boolean = {
-    if (pruneNonOptimal) {
+  def shouldEnterState(curr: State): Boolean = {
+    val splitToken = toks(curr.splits.length)
+    val insideOptimizationZone =
+      curr.policy.noDequeue || isInsideNoOptZone(splitToken)
+    def hasBestSolution = !pruneNonOptimal || insideOptimizationZone || {
       val splitToken = toks(curr.splits.length)
       // TODO(olafur) document why/how this optimization works.
-      val hasBest = best.get(splitToken.left).exists(_.alwaysBetter(curr))
-      hasBest
-    } else {
-      false
+      !best.get(splitToken.left).exists(_.alwaysBetter(curr))
     }
+   hasBestSolution
+  }
+
+  def getLeftLeft(curr: State): Token = {
+    toks(Math.max(0, curr.splits.length - 1)).left
+  }
+
+  def shouldRecurseOnBlock(curr: State, stop: Token) = {
+    val leftLeft = getLeftLeft(curr)
+    val leftLeftOwner = owners(hash(leftLeft))
+    val splitToken = toks(curr.splits.length)
+    recurseOnBlocks &&
+      isInsideNoOptZone(splitToken) &&
+      leftLeft.isInstanceOf[`{`] &&
+      matchingParentheses(hash(leftLeft)) != stop &&
+      {
+        // Block must span at least 3 lines to be worth recursing.
+        val close = matchingParentheses(hash(leftLeft))
+        // TODO(olafur) magic number
+        close.start - leftLeft.end > style.maxColumn * 3
+      } &&
+      extractStatementsIfAny(leftLeftOwner).nonEmpty
   }
 
   def provided(formatToken: FormatToken): Split = {
@@ -73,113 +120,128 @@ class BestFirstSearch(style: ScalaStyle, tree: Tree, range: Set[Range])
     result
   }
 
+  def stateColumnKey(state: State): Int = {
+    state.column << 8 | state.indentation
+  }
+
+  def hasReachedEof(state: State): Boolean = {
+    explored > maxVisitStates || state.splits.length == toks.length
+  }
+
+  val memo = mutable.Map.empty[(Int, StateHash), State]
+
+  def shortestPathMemo(start: State, stop: Token, depth: Int): State = {
+    memo.getOrElseUpdate((start.splits.length, stateColumnKey(start)), {
+      shortestPath(start, stop, depth)
+    })
+  }
+
   /**
-    *
     * Runs best first search to find lowest penalty split.
     */
-  def shortestPath(owner: Tree, start: State): State = {
-    if (style.debug) {
-      Debug.visit(owner)
-    }
-    logger.trace(
-      s"""${ start.indentation } ${start.splits.takeRight(3)}
-         |${log(owner)}
-         |FORMAT:
-         |${ State. reconstructPath(toks, start.splits, style) }""".stripMargin)
-
+  def shortestPath(start: State, stop: Token, depth: Int = 0,
+                   maxCost: Int = Integer.MAX_VALUE): State = {
     val Q = new mutable.PriorityQueue[State]()
     var result = start
     Q += start
+    // TODO(olafur) this while loop is waaaaaaaaaaaaay tooo big.
     while (Q.nonEmpty) {
       val curr = Q.dequeue()
       explored += 1
-      if (explored % 1000 == 0 && style.debug) {
+      if (explored % 10000 == 0 && style.debug) {
         logger.debug(s"Explored $explored")
       }
-      val i = curr.splits.length
-      if (explored > maxVisitStates ||
-        i == toks.length ||
-        !childOf(owners(hash(toks(i).right)), owner)) {
+      if (hasReachedEof(curr) ||
+        toks(curr.splits.length).left == stop) {
         result = curr
-        logger.trace(
-          s"""
-             |${log(owner, tokensOnly = true)}
-             |${toks.length}
-             |${curr.splits.length}
-             |Q.size: ${Q.size}
-             |${curr.splits.length}
-           """.stripMargin)
         Q.dequeueAll
         Unit
-      }
-      else {
-        val splitToken = toks(i)
-        val noOpt = curr.policy.noDequeue || isInsideNoOptZone(splitToken)
-        if (noOpt || !pruneOK(curr)) {
-          if (curr.splits.length > deepestYet.splits.length) {
-            deepestYet = curr
-            // Only dequeue if this solution is the deepest solution so far.
-          }
-          if (dequeOnNewStatements &&
-            !curr.policy.noDequeue &&
-            statementStarts.contains(hash(splitToken.left)) &&
-            curr.splits.last.modification.isNewline ) {
-            logger.trace(Q.map(s => mkString(s.splits)).mkString("\n\n"))
-            logger.trace(mkString(curr.splits))
-            Q.dequeueAll
-          }
-          if (style.debug) {
-            Debug.visit(splitToken)
-          }
-          if (Q.nonEmpty) {
-            logger.trace(
-              s"""
-                 |visit=$splitToken
-                 |lastSplit=${curr.splits.last}
-                 |cost=${curr.cost}
-                 |Q.size=${Q.size}
-                 |""".stripMargin)
-          }
+      } else if (shouldEnterState(curr)) {
+        val splitToken = toks(curr.splits.length)
+        if (curr.splits.length > deepestYet.splits.length) {
+          deepestYet = curr
+          // Only dequeue if this solution is the deepest solution so far.
+        }
+
+        if (dequeOnNewStatements &&
+          (depth > 0 || !isInsideNoOptZone(splitToken)) &&
+          statementStarts.contains(hash(splitToken.left)) &&
+          curr.splits.last.modification.isNewline ) {
+          Q.dequeueAll
+        }
+
+        if (style.debug) {
+          Debug.visit(splitToken)
+        }
+        if (shouldRecurseOnBlock(curr, stop)) {
+          val nextState = shortestPathMemo(
+            curr,
+            matchingParentheses(hash(getLeftLeft(curr))),
+            depth = depth + 1)
+          Q.enqueue(nextState)
+        } else {
+
           val splits: Seq[Split] =
             if (splitToken.inside(range))
               router.getSplitsMemo(splitToken)
             else List(provided(splitToken))
 
-          val actualSplit = curr.policy
-            .execute(Decision(splitToken, splits)).splits
-          if (splits.length != actualSplit.length) {
-            val activePolicies = curr.policy.policies
-              .filterNot(_.expire < splitToken.left.end)
-            logger.trace(s"${activePolicies.mkString(", ")} killed $splits")
+
+
+          val actualSplit = {
+            curr.policy
+              .execute(Decision(splitToken, splits)).splits
+              .filter(!_.ignoreIf)
+              .sortBy(_.cost)
           }
-          actualSplit.withFilter(!_.ignoreIf).foreach { split =>
+          var optimalNotFound = true
+          actualSplit.foreach { split =>
             val nextState = curr.next(style, split, splitToken)
             // TODO(olafur) convince myself this is safe.
             if (split.modification.isNewline) {
-              best += splitToken.left -> nextState
+              best.update(splitToken.left, nextState)
             }
-            Q.enqueue(nextState)
+            if (style.debug) {
+              Debug.enqueued(split)
+            }
+            split.optimalAt match {
+              case Some(token) if actualSplit.length > 1 && split.cost == 0 =>
+                val nextNextState = shortestPath(nextState, token, depth, maxCost = 0)
+                if (toks(nextNextState.splits.length).left == token) {
+                  optimalNotFound = false
+                  Q.enqueue(nextNextState)
+                } else if (nextState.cost - curr.cost <= maxCost) {
+                  // TODO(olafur) DRY. This solution can still be optimal.
+                  Q.enqueue(nextState)
+                }
+              case _ if optimalNotFound && nextState.cost - curr.cost <= maxCost =>
+                Q.enqueue(nextState)
+              case _ => // Other split was optimal
+            }
           }
-          Unit
         }
+        Unit
       }
     }
     result
   }
 
   def formatTree(): String = {
-    logger.trace(log(tree, tokensOnly = true))
-    val state = shortestPath(tree, State.start)
+    var state = shortestPath(State.start, tree.tokens.last)
     if (state.splits.length != toks.length) {
-      val msg =
-        s"""UNABLE TO FORMAT,
-            |state.length=${state.splits.length}
-            |toks.length=${toks.length}
-            |deepestYet.length=${deepestYet.splits.length}
-            |Output:
-            |${mkString(deepestYet.splits)}
-            |""".stripMargin
-      throw Error.CantFormatFile(msg)
+      if (style.debug) {
+        state = deepestYet
+      } else {
+        val msg =
+          s"""UNABLE TO FORMAT,
+              |state.length=${state.splits.length}
+              |toks.length=${toks.length}
+              |deepestYet.length=${deepestYet.splits.length}
+              |Output:
+              |${mkString(deepestYet.splits)}
+              |""".stripMargin
+        throw Error.CantFormatFile(msg)
+      }
     }
     if (style.debug) {
       Debug.explored += explored
@@ -200,9 +262,42 @@ class BestFirstSearch(style: ScalaStyle, tree: Tree, range: Set[Range])
     sb.toString()
   }
 
+  def noOptimizationZones(tree: Tree): Set[Token] = {
+    val result = new mutable.SetBuilder[Token, Set[Token]](Set.empty[Token])
+    var inside = false
+    var expire = tree.tokens.head
+    tree.tokens.foreach {
+      case t: `(` if owners(hash(t)).isInstanceOf[Term.Apply] && !inside =>
+        inside = true
+        expire = matchingParentheses(hash(t))
+      case x if x == expire =>
+        inside = false
+      case x if inside =>
+        result += x
+      case _ =>
+    }
+    result.result()
+  }
+
 }
 
 object BestFirstSearch extends ScalaFmtLogger {
+
+  def extractStatementsIfAny(tree: Tree): Seq[Tree] = tree match {
+    case b: Term.Block => b.stats
+    case t: Pkg => t.stats
+    case t: Term.For => t.enums
+    case t: Term.ForYield => t.enums
+    case t: Term.Match => t.cases
+    case t: Term.PartialFunction => t.cases
+    case t: Term.TryWithCases => t.catchp
+    case t: Type.Compound => t.refinement
+    case t: scala.meta.internal.ast.Source => t.stats
+    case t: Template if t.stats.isDefined => t.stats.get
+    case _ => Seq.empty[Tree]
+  }
+
+
   def getStatementStarts(tree: Tree): Map[TokenHash, Tree] = {
     val ret =
       new mutable.MapBuilder[TokenHash, Tree, Map[TokenHash, Tree]](Map[TokenHash, Tree]())
@@ -233,7 +328,6 @@ object BestFirstSearch extends ScalaFmtLogger {
 
     def loop(x: Tree): Unit = {
       x match {
-        case b: Term.Block => addAll(b.stats)
         case t: Defn.Class => addDefn[`class `](t.mods, t)
         case t: Defn.Def => addDefn[`def`](t.mods, t)
         case t: Defn.Object => addDefn[`object`](t.mods, t)
@@ -241,17 +335,8 @@ object BestFirstSearch extends ScalaFmtLogger {
         case t: Defn.Type => addDefn[`type`](t.mods, t)
         case t: Defn.Val => addDefn[`val`](t.mods, t)
         case t: Defn.Var => addDefn[`var`](t.mods, t)
-        case t: Pkg => addAll(t.stats)
-        case t: Term.For => addAll(t.enums)
-        case t: Term.ForYield => addAll(t.enums)
-        case t: Term.Match => addAll(t.cases)
-        case t: Term.PartialFunction => addAll(t.cases)
-        case t: Term.TryWithCases => addAll(t.catchp)
-        case t: Type.Compound => addAll(t.refinement)
-        case t: scala.meta.internal.ast.Source => addAll(t.stats)
-        case t: Template if t.stats.isDefined =>
-          addAll(t.stats.get)
-        case _ => // Nothing
+        case t => // Nothing
+          addAll(extractStatementsIfAny(t))
       }
       x.children.foreach(loop)
     }
@@ -304,18 +389,6 @@ object BestFirstSearch extends ScalaFmtLogger {
         result += hash(tok) -> x
       }
       x.children.foreach(loop)
-    }
-    loop(tree)
-    result.result()
-  }
-
-  def noOptimizationZones(tree: Tree): Set[Token] = {
-    val result = new mutable.SetBuilder[Token, Set[Token]](Set.empty[Token])
-    def loop(x: Tree): Unit = x match {
-      case t: Term.Apply =>
-        result ++= t.tokens
-      case _ =>
-        x.children.foreach(loop)
     }
     loop(tree)
     result.result()