@olafurpg 2016-02-26
A glimpse of how much state got eliminated.

<img width="466" alt="screen shot 2016-02-26 at 17 38 00" src="https://cloud.githubusercontent.com/assets/1408093/13358462/b8e079d8-dcaf-11e5-9ed4-a26491d0bb6d.png">